move public functions to index.js
